# Performance improvements using iterator aggregates