--- a/src/Contract/Operation/Applyable.php
+++ b/src/Contract/Operation/Applyable.php
@@ -9,7 +9,6 @@ declare(strict_types=1);
 
 namespace loophp\collection\Contract\Operation;
 
-use Iterator;
 use loophp\collection\Contract\Collection;
 
 /**
@@ -25,7 +24,7 @@ interface Applyable
      *
      * @see https://loophp-collection.readthedocs.io/en/stable/pages/api.html#apply
      *
-     * @param callable(T=, TKey=, Iterator<TKey, T>=): bool ...$callbacks
+     * @param callable(T=, TKey=, iterable<TKey, T>=): bool ...$callbacks
      *
      * @return Collection<TKey, T>
      */