--- a/src/Contract/Operation/Everyable.php
+++ b/src/Contract/Operation/Everyable.php
@@ -9,8 +9,6 @@ declare(strict_types=1);
 
 namespace loophp\collection\Contract\Operation;
 
-use Iterator;
-
 /**
  * @template TKey
  * @template T
@@ -22,7 +20,7 @@ interface Everyable
      *
      * @see https://loophp-collection.readthedocs.io/en/stable/pages/api.html#every
      *
-     * @param callable(T=, TKey=, Iterator<TKey, T>=): bool ...$callbacks
+     * @param callable(T=, TKey=, iterable<TKey, T>=): bool ...$callbacks
      */
     public function every(callable ...$callbacks): bool;
 }