--- a/src/Contract/Operation/Hasable.php
+++ b/src/Contract/Operation/Hasable.php
@@ -9,8 +9,6 @@ declare(strict_types=1);
 
 namespace loophp\collection\Contract\Operation;
 
-use Iterator;
-
 /**
  * @template TKey
  * @template T
@@ -22,7 +20,7 @@ interface Hasable
      *
      * @see https://loophp-collection.readthedocs.io/en/stable/pages/api.html#has
      *
-     * @param callable(T=, TKey=, Iterator<TKey, T>=): T ...$callbacks
+     * @param callable(T=, TKey=, iterable<TKey, T>=): T ...$callbacks
      */
     public function has(callable ...$callbacks): bool;
 }