--- a/src/Contract/Operation/Partitionable.php
+++ b/src/Contract/Operation/Partitionable.php
@@ -9,7 +9,6 @@ declare(strict_types=1);
 
 namespace loophp\collection\Contract\Operation;
 
-use Iterator;
 use loophp\collection\Contract\Collection;
 
 /**
@@ -26,7 +25,7 @@ interface Partitionable
      *
      * @see https://loophp-collection.readthedocs.io/en/stable/pages/api.html#partition
      *
-     * @param callable(T=, TKey=, Iterator<TKey, T>=): bool ...$callbacks
+     * @param callable(T=, TKey=, iterable<TKey, T>=): bool ...$callbacks
      *
      * @return Collection<int, Collection<TKey, T>>
      */