--- a/src/Contract/Operation/Rejectable.php
+++ b/src/Contract/Operation/Rejectable.php
@@ -9,7 +9,6 @@ declare(strict_types=1);
 
 namespace loophp\collection\Contract\Operation;
 
-use Iterator;
 use loophp\collection\Contract\Collection;
 
 /**
@@ -23,7 +22,7 @@ interface Rejectable
      *
      * @see https://loophp-collection.readthedocs.io/en/stable/pages/api.html#reject
      *
-     * @param callable(T=, TKey=, Iterator<TKey, T>=): bool ...$callbacks
+     * @param callable(T=, TKey=, iterable<TKey, T>=): bool ...$callbacks
      *
      * @return Collection<TKey, T>
      */