--- a/src/Contract/Operation/Spanable.php
+++ b/src/Contract/Operation/Spanable.php
@@ -9,7 +9,6 @@ declare(strict_types=1);
 
 namespace loophp\collection\Contract\Operation;
 
-use Iterator;
 use loophp\collection\Contract\Collection;
 
 /**
@@ -27,7 +26,7 @@ interface Spanable
      *
      * @see https://loophp-collection.readthedocs.io/en/stable/pages/api.html#span
      *
-     * @param callable(T=, TKey=, Iterator<TKey, T>=): bool ...$callbacks
+     * @param callable(T=, TKey=, iterable<TKey, T>=): bool ...$callbacks
      *
      * @return Collection<int, Collection<TKey, T>>
      */