--- a/src/Contract/Operation/TakeWhileable.php
+++ b/src/Contract/Operation/TakeWhileable.php
@@ -9,7 +9,6 @@ declare(strict_types=1);
 
 namespace loophp\collection\Contract\Operation;
 
-use Iterator;
 use loophp\collection\Contract\Collection;
 
 /**
@@ -24,7 +23,7 @@ interface TakeWhileable
      *
      * @see https://loophp-collection.readthedocs.io/en/stable/pages/api.html#takewhile
      *
-     * @param callable(T=, TKey=, Iterator<TKey, T>=): bool ...$callbacks
+     * @param callable(T=, TKey=, iterable<TKey, T>=): bool ...$callbacks
      *
      * @return Collection<TKey, T>
      */