--- a/src/Operation/Coalesce.php
+++ b/src/Operation/Coalesce.php
@@ -11,7 +11,6 @@ namespace loophp\collection\Operation;
 
 use Closure;
 use Generator;
-use Iterator;
 
 /**
  * @immutable
@@ -24,11 +23,11 @@ final class Coalesce extends AbstractOperation
     /**
      * @pure
      *
-     * @return Closure(Iterator<TKey, T>): Generator<TKey, T>
+     * @return Closure(iterable<TKey, T>): Generator<TKey, T>
      */
     public function __invoke(): Closure
     {
-        /** @var Closure(Iterator<TKey, T>): Generator<TKey, T> $pipe */
+        /** @var Closure(iterable<TKey, T>): Generator<TKey, T> $pipe */
         $pipe = Pipe::of()(
             Compact::of()(),
             Head::of(),