… composer.json
… src/Collection.php
… src/Operation/Combinate.php
… src/Operation/Combine.php
… src/Operation/Compact.php
… src/Operation/Contains.php
… src/Operation/Current.php
… src/Operation/Cycle.php
… src/Operation/Diff.php
… src/Operation/DiffKeys.php
… src/Operation/Distinct.php
… src/Operation/DropWhile.php
… src/Operation/Dump.php
… src/Operation/Duplicate.php
… src/Operation/Equals.php
… src/Operation/Every.php
… src/Operation/Explode.php
… src/Operation/Falsy.php
… src/Operation/Filter.php
… src/Operation/Find.php
… src/Operation/First.php
… src/Operation/FlatMap.php
… src/Operation/Flatten.php
… src/Operation/Flip.php
… src/Operation/FoldLeft.php
… src/Operation/FoldLeft1.php
… src/Operation/FoldRight.php
… src/Operation/FoldRight1.php
… src/Operation/Forget.php
… src/Operation/Frequency.php