This PR:

- [x] Replaces `IterableIterator` with `IterableIteratorAggregate` (4 occurences)
- [x] Changes the signature of a couple of operations. 
- [x] Update all the annotations where an `Iterator` is needed while now an `iterable` is enough.
- [x] Use `yield from` - align everything
- [x] Get rid of `ClosureIterator` and replace it with `ClosureIteratorAggregate`

Based on the benchmarks I did in `loophp/iterators`, it seems to be definitely, and it make sense.
`IteratorAggregates` are usually faster because the state of the `Iterator` is not built in _userland_ but directly handled by PHP internaly.


```
$ ./vendor/bin/phpbench run --profile=local --group=local --report=aggregate --stop-on-error tests/benchmarks/IterableIteratorAggregateBench.php tests/benchmarks/IterableIteratorBench.php
PHPBench (1.2.3) running benchmarks...
with configuration file: /home/pol/dev/git/iterators/phpbench.json
with PHP version 8.1.1, xdebug ❌, opcache ✔

\benchmarks\loophp\iterators\IterableIteratorAggregateBench

    bench # IterableIteratorAggregate.......I0 - Mo1.963ms (±0.00%)
    bench # IterableIteratorAggregate.......R1 I7 - Mo0.937ms (±1.97%)
    bench # IterableIteratorAggregate.......R1 I25 - Mo0.927ms (±1.60%)

\benchmarks\loophp\iterators\IterableIteratorBench

    bench # IterableIterator................R1 I0 - Mo2.188ms (±0.00%)
    bench # IterableIterator................R1 I6 - Mo1.765ms (±1.80%)
    bench # IterableIterator................R4 I49 - Mo1.741ms (±2.25%)

Subjects: 2, Assertions: 0, Failures: 0, Errors: 0
+--------------------------------+---------+---------------------------+------+-----+----------+---------+--------+
| benchmark                      | subject | set                       | revs | its | mem_peak | mode    | rstdev |
+--------------------------------+---------+---------------------------+------+-----+----------+---------+--------+
| IterableIteratorAggregateBench | bench   | IterableIteratorAggregate | 10   | 1   | 3.695mb  | 1.963ms | ±0.00% |
| IterableIteratorAggregateBench | bench   | IterableIteratorAggregate | 10   | 10  | 3.695mb  | 0.937ms | ±1.97% |
| IterableIteratorAggregateBench | bench   | IterableIteratorAggregate | 10   | 50  | 3.695mb  | 0.927ms | ±1.60% |
| IterableIteratorBench          | bench   | IterableIterator          | 10   | 1   | 3.695mb  | 2.188ms | ±0.00% |
| IterableIteratorBench          | bench   | IterableIterator          | 10   | 10  | 3.695mb  | 1.765ms | ±1.80% |
| IterableIteratorBench          | bench   | IterableIterator          | 10   | 50  | 3.695mb  | 1.741ms | ±2.25% |
+--------------------------------+---------+---------------------------+------+-----+----------+---------+--------+
```

@AlexandruGG WDYT of this? Are you ok with this?